refactor: extract roll table orchestration and split MSW mutation helpers
