+++ src/msw/api/businesses/[business-id]/bank-transactions/bulk-match-or-categorize/post.ts
@@ -2,7 +2,8 @@ import { Schema } from 'effect'
 
 import { BulkMatchOrCategorizeRequestSchema } from '@schemas/bankTransactions/bulkMatchOrCategorize'
 
-import { applyCategoryUpdate, applyConfirmedMatch } from '@msw/api/businesses/[business-id]/bank-transactions/applyCategoryUpdate'
+import { applyCategoryUpdate } from '@msw/api/businesses/[business-id]/bank-transactions/applyCategoryUpdate'
+import { applyConfirmedMatch } from '@msw/api/businesses/[business-id]/bank-transactions/applyConfirmedMatch'
 import { bankTransactionStore, findOrSeedBankTransaction } from '@msw/api/businesses/[business-id]/bank-transactions/store'
 import { apiData } from '@msw/utils/apiResponse'
 import { createMockEndpoint } from '@msw/utils/createMockEndpoint'
